requirements.txt
flask-sqlalchemy==3.1.1
flask-mysqldb==2.0.0
matplotlib==3.8.0